requirements.txt
python-dotenv
openai
pandas
streamlit > 1.43